Add files via upload
Full Matlab code workflow chain for the trajectory tracking and statistical analysis of rigid and healthy RBCs in bifurcations. 
This code is associated with the manuscript
Impact of Red Blood Cell Rigidity on in vivo Flow Dynamics and Lingering in Bifurcations
by Rashidi et al. 2025 .

This code uses Shape Language Modelling and FastPeakFind from the Matlab repositories.